Statistics: Fix protocol buffers edge case with empty stats
Attempting to serialize a `Statistic` with zero values added would
result in a vague complaint: ""JsonObjectWriter was not fully closed".

Some sleuthing pintpointed this to `std::nan` values returned by
`mean()` and `pstdev()`, Which indeed is not a good value to attempt
to pass to `set_nanos` if proto duration fields.

Signed-off-by: Otto van der Schaaf <oschaaf@we-amp.com>