Ssl: Add MinimalTransportSocketFactoryContext - WIP
Minimal implementation of TransportSocketFactoryContext.
Will assert when attempting to get a dependency that is not set up.

Manual testing shows that the basics for ssl (and h2) are working.
Didn't go in a straight line from a->b, this needs cleanup.

https/h2 test was disabled because of it needing thead local storage.

Signed-off-by: Otto van der Schaaf <oschaaf@we-amp.com>